fix methods which calls layout class
