+++ src/pyedb/dotnet/edb_core/cell/layout.py
@@ -33,13 +33,26 @@ from pyedb.dotnet.edb_core.edb_data.primitives_data import (
     EdbRectangle,
     EdbText,
 )
-from pyedb.dotnet.edb_core.edb_data.sources import PinGroup
+
 from pyedb.dotnet.edb_core.general import convert_py_list_to_net_list
+from pyedb.dotnet.edb_core.cell.hierarchy.component import EDBComponent
 from pyedb.dotnet.edb_core.utilities.obj_base import ObjBase
+from pyedb.dotnet.edb_core.cell.terminal.pingroup_terminal import PinGroupTerminal
+from pyedb.dotnet.edb_core.cell.terminal.point_terminal import PointTerminal
+from pyedb.dotnet.edb_core.cell.terminal.padstack_instance_terminal import PadstackInstanceTerminal
+from pyedb.dotnet.edb_core.cell.terminal.edge_terminal import EdgeTerminal
+from pyedb.dotnet.edb_core.cell.terminal.bundle_terminal import BundleTerminal
+from pyedb.dotnet.edb_core.edb_data.nets_data import (
+    EDBDifferentialPairData,
+    EDBExtendedNetData,
+    EDBNetClassData,
+)
+from pyedb.dotnet.edb_core.edb_data.sources import PinGroup
 
 
-class LayoutToRemove:
-    # todo refactor
+class Layout(ObjBase):
+    def __init__(self, pedb, edb_object):
+        super().__init__(pedb, edb_object)
 
     @property
     def cell(self):
@@ -60,68 +73,8 @@ class LayoutToRemove:
         self._edb_object.SetLayerCollection(layer_collection)
 
     @property
-    def terminals(self):
-        """:obj:`list` of :class:`Terminal <ansys.edb.terminal.Terminal>` : List of all the terminals in this layout.
-
-        Read-Only.
-        """
-        return list(self._edb_object.Terminals)
-
-    @property
-    def groups(self):
-        """:obj:`list` of :class:`Group <ansys.edb.hierarchy.Group>` : List of all the groups in this layout.
-
-        Read-Only.
-        """
-        return list(self._edb_object.Groups)
-
-    @property
-    def net_classes(self):
-        """:obj:`list` of :class:`NetClass <ansys.edb.net.NetClass>` : List of all the netclasses in this layout.
-
-        Read-Only.
-        """
-        return list(self._edb_object.NetClasses)
-
-    @property
-    def differential_pairs(self):
-        """:obj:`list` of :class:`DifferentialPair <ansys.edb.net.DifferentialPair>` : List of all the differential \
-         pairs in this layout.
-
-        Read-Only.
-        """
-        return list(self._edb_object.DifferentialPairs)
-
-    @property
-    def pin_groups(self):
-        """:obj:`list` of :class:`PinGroup <ansys.edb.hierarchy.PinGroup>` : List of all the pin groups in this \
-        layout.
-
-        Read-Only.
-        """
-        return list(self._edb_object.PinGroups)
-
-    @property
-    def voltage_regulators(self):
-        """:obj:`list` of :class:`VoltageRegulator <ansys.edb.hierarchy.VoltageRegulator>` : List of all the voltage \
-         regulators in this layout.
-
-        Read-Only.
-        """
-        return list(self._edb_object.VoltageRegulators)
-
-    @property
-    def extended_nets(self):
-        """
-        Get the list of extended nets in the layout. Read-Only.
-
-        Returns
-        -------
-        List[:class:`ExtendedNet <ansys.edb.net.ExtendedNet>`]
-            A list of extended nets.
-
-        """
-        return list(self._edb_object.ExtendedNets)
+    def _edb(self):
+        return self._pedb.edb_api
 
     def expanded_extent(self, nets, extent, expansion_factor, expansion_unitless, use_round_corner, num_increments):
         """Get an expanded polygon for the Nets collection.
@@ -191,14 +144,28 @@ class LayoutToRemove:
     def fixed_zone_primitive(self, value):
         self._edb_object.SetFixedZonePrimitives(value)
 
-
-class Layout(ObjBase, LayoutToRemove):
-    def __init__(self, pedb, edb_object):
-        super().__init__(pedb, edb_object)
-
     @property
-    def _edb(self):
-        return self._pedb.edb_api
+    def terminals(self):
+        """Get terminals belonging to active layout.
+
+        Returns
+        -------
+        Terminal dictionary : Dict[str, pyedb.dotnet.edb_core.edb_data.terminals.Terminal]
+        """
+        temp = []
+        for i in list(self._edb_object.Terminals):
+            terminal_type = i.ToString().split(".")[-1]
+            if terminal_type == "PinGroupTerminal":
+                temp.append(PinGroupTerminal(self._pedb, i))
+            elif terminal_type == "PadstackInstanceTerminal":
+                temp.append(PadstackInstanceTerminal(self._pedb, i))
+            elif terminal_type == "EdgeTerminal":
+                temp.append(EdgeTerminal(self._pedb, i))
+            elif terminal_type == "BundleTerminal":
+                temp.append(BundleTerminal(self._pedb, i))
+            elif terminal_type == "PointTerminal":
+                temp.append(PointTerminal(self._pedb, i))
+        return temp
 
     @property
     def cell_instances(self):
@@ -268,24 +235,34 @@ class Layout(ObjBase, LayoutToRemove):
         list :
             List of bondwires.
         """
-        return [
-            Bondwire(self._pedb, i)
-            for i in self._edb_object.Primitives
-            if i.GetPrimitiveType().ToString() == "Bondwire"
-        ]
+        return [i for i in self.primitives if i.primitive_type == "bondwire"]
 
     @property
-    def board_bend_defs(self):
-        """:obj:`list` of :class:`BoardBendDef <ansys.edb.primitive.BoardBendDef>` : List of all the board bend \
-        definitions in this layout.
+    def groups(self):
+        temp = []
+        for i in list(self._edb_object.Groups):
+            group_type = i.ToString().split(".")[-1].lower()
+            if group_type == "component":
+                temp.append(EDBComponent(self._pedb, i))
+            else:
+                pass
+        return temp
 
-        Read-Only.
-        """
-        return list(self._edb_object.GetBoardBendDefs())
+    @property
+    def pin_groups(self):
+        return [PinGroup(pedb=self._pedb, edb_pin_group=i, name=i.GetName()) for i in self._edb_object.PinGroups]
 
-    def synchronize_bend_manager(self):
-        """Synchronize bend manager."""
-        self._edb_object.SynchronizeBendManager()
+    @property
+    def net_classes(self):
+        return [EDBNetClassData(self._pedb, i) for i in list(self._edb_object.NetClasses)]
+
+    @property
+    def extended_nets(self):
+        return [EDBExtendedNetData(self._pedb, i) for i in self._edb_object.ExtendedNets]
+
+    @property
+    def differential_pairs(self):
+        return [EDBDifferentialPairData(self._pedb, i) for i in list(self._edb_object.DifferentialPairs)]
 
     @property
     def padstack_instances(self):
@@ -310,11 +287,12 @@ class Layout(ObjBase, LayoutToRemove):
         """
         obj = self._pedb._edb.Cell.Connectable.FindById(self._edb_object, value)
         if obj.GetObjType().ToString() == "PadstackInstance":
-            return EDBPadstackInstance(obj, self._pedb)
+            return EDBPadstackInstance(obj, self._pedb) if obj is not None else None
 
     def find_net_by_name(self, value: str):
         obj = self._pedb._edb.Cell.Net.FindByName(self._edb_object, value)
-        return EDBNetsData(obj, self._pedb)
+        return EDBNetsData(obj, self._pedb) if obj is not None else None
 
-        obj = self._edb.cell.hierarchy.pin_group.Create(self._edb_object, name, convert_py_list_to_net_list(pins))
-        return PinGroup(name, obj, self._pedb)
+    def find_component_by_name(self, value: str):
+        obj = self._pedb._edb.Cell.Hierarchy.Component.FindByName(self._edb_object, value)
+        return EDBComponent(self._pedb, obj) if obj is not None else None
