+++ src/pyedb/dotnet/edb_core/cell/terminal/terminal.py
@@ -160,24 +160,6 @@ class Terminal(Connectable):
         self._port_post_processing_prop = ppp
 
     @property
-    def name(self):
-        """Port Name.
-
-        Returns
-        -------
-        str
-        """
-        return self._edb_object.GetName()
-
-    @name.setter
-    def name(self, value):
-        if isinstance(value, str):
-            if not any(port for port in list(self._pedb.excitations.keys()) if port == value):
-                self._edb_object.SetName(value)
-            else:
-                self._pedb.logger.warning("An existing port already has this same name. A port name must be unique.")
-
-    @property
     def net_name(self):
         """Net name.
